// src/model/user.model.ts
export interface UserModel {
    // --- Champs Google ---
    googleId?: string;
    googleDisplayName?: string;
    googleEmail?: string;
    googlePhoto?: string;

    // --- Champs GitHub ---
    githubId?: string;
    githubUsername?: string;
    githubEmail?: string;
    githubPhoto?: string;
}